// web-ANGULAR/dfps-web-impact/src/app/modules/case/model/homeAssessment.ts
export interface DisplayHomeAssessmentListAdd {
    stageId: number;
    eventMap: EventMap;
    pageMode: string;
    stageName: string;
}

export class EventMap {
    cpsHomeAssessment: EventType;
    contractHomeAssessment: EventType;
    kinshipSafetyEvaluation: EventType;
}

export class EventType {
    eventType: string;
}

export class FormParamsDto {
    action: string;
    approvalFlag: string;
    attachmentFormName: string;
    bFormContact: string;
    bPreview: string;
    bResend: string;
    ccEmailAddress: string;
    cdContactMethod: string;
    cdContactOthers: string;
    cdContactPurpose: string;
    cdRelation: string;
    cdStage: string;
    checkForNewMode: string;
    checkStage: string;
    cpsContractAddendum: string;
    deleteDocument: string;
    displayName: string;
    displayOnly: string;
    docExists: string;
    docType: string;
    dtContactOccurred: string;
    dtTimestamp: string;
    emailContent: string;
    emailFrom: string;
    emailNotify: string;
    emailTitle: string;
    emailToList: string;
    eventType: string;
    formName: string;
    hdnCalledFromJavascript: string;
    hdnFlagcontractaddendum: string;
    hdnFlagcpsaddendum: string;
    hiddenField: string;
    hideInWidget: string;
    id: string;
    idCase: string;
    idCreatedPerson: string;
    idEvent: string;
    idEventReq: string;
    idEventTable: string;
    idFamilyPlanEvalEvent: string;
    idFamilyPlanEvent: string;
    idFormReferral: string;
    idFsnaEvent: string;
    idICPCTransmittal: string;
    idLastUpdatePerson: string;
    idLegalNoticeDtl: string;
    idLegalNoticeRecpnt: string;
    idMedCons: string;
    idPPTEvent: string;
    idPPTPart: string;
    idPerson: string;
    idPersonTable: string;
    idPrmryPrntCrgvr: string;
    idReferral: string;
    idResource: string;
    idResourceTable: string;
    idRiskAsmntLkp: string;
    idSSCCPlcmtHeader: string;
    idScndryPrntCrgvr: string;
    idStage: string;
    idSvcPlnEvalEvent: string;
    idTransmittal: string;
    indAddOrModify: string;
    level1Tab: string;
    level2Tab: string;
    level3Tab: string;
    modeOfPage: string;
    nCase: string;
    nEvent: string;
    name: string;
    nbrVersion: string;
    nmStage: string;
    onClick: string;
    pArClosureReason: string;
    pCase: string;
    pCdNotifStat: string;
    pCdStage: string;
    pChild: string;
    pCrimHist: string;
    pDtCPSInvstDtlBegun: string;
    pDtLicngInvstDtlBegun: string;
    pDtSampleFrom: string;
    pDtSampleTo: string;
    pEvent: string;
    pFormName: string;
    pIcpcRequest: string;
    pMedCons: string;
    pParameter: string;
    pPerson: string;
    pPersonCitizenship: string;
    pPersonId: string;
    pPptEvent: string;
    pRecCheck: string;
    pRecordsCheckNotif: string;
    pRegion: string;
    pRejection: string;
    pStage: string;
    pStageReasonClosed: string;
    pSzCdEventType: string;
    pSzCdStageType: string;
    pWorker: string;
    postInSameWindow: string;
    promptNoDocument: string;
    promptSavePage: string;
    protectDocument: string;
    renderFormat: string;
    sCase: string;
    sEvent: string;
    sIdRecCheck: string;
    sIdTletsCheck: string;
    sRecordsCheckNotif: string;
    sStage: string;
    sTimestamp: string;
    selCdSearchType: string;
    selSzCdStageReasonClosed: string;
    selectedEpisodes: string;
    switchLanguage: string;
    taskCD: string;
    timestamp: string;
    tmScrTmCntct: string;
    txtNmPersonFull: string;
    userId: string;
    userIdPerson: string;
    userName: string;
    windowName: string;
    xpandContactSearch: string;
    xpandGlobalData: string;
    pResource: string;
    newTimestamp: string;
}